+++ src/components/Organisms/StoreSelectAutocomplete/MockApiServerResponseData.ts
